refactor(select): rewrite to agreed format
